Fix buffer allocation type in RenderBundleEncoderDescriptor
Correct the calculation for buffer allocation size by ensuring type consistency. Additionally, added import for WGPUInstanceExtras in InstanceDescriptor for extended functionalities.

+++ wgpu4k/src/jvmMain/kotlin/io.ygdrasil.wgpu/mapper/RenderBundleEncoderDescriptor.kt
@@ -12,7 +12,7 @@ fun MemoryAllocator.map(input: RenderBundleEncoderDescriptor): WGPURenderBundleE
         if (input.colorFormats.isNotEmpty()) {
             output.colorFormatCount = input.colorFormats.size.toULong()
 
-            output.colorFormats = allocateBuffer(Int.SIZE_BYTES * output.colorFormatCount)
+            output.colorFormats = allocateBuffer(Int.SIZE_BYTES.toULong() * output.colorFormatCount)
                 .also { it.writeInts(input.colorFormats.map { it.value }.toIntArray()) }
                 .let { ArrayHolder(it.handler) }
         }
